# [logging] more concise logging for resource optimization